Add magic values to packets for DOS protection
It was trivial to mount a DOS attack by simply sending a valid IP packet
to the server and the server will henceforth send outbound UDP packets
to the adversary instead of the intended remote host. We add a magic value
that is prepended to every UDP message that is known only to the client
and the server. The server will only switch to new a IP address if they
had sent a UDP message with a valid magic value.

This scheme does not protect against replay attack, but that is okay since
any adversary that could intercept UDP packets already has the power to
disrupt communication between the client and server.